+++ Src/LanguageEnglish.lua
@@ -65,7 +65,7 @@ function englishModule:Translations()
     ["options.short.Carrot"]                          = "Warn about mount items",
     ["options.short.DeathBlock"]                      = "Pause if someone is dead",
     ["options.short.DontUseConsumables"]              = "Dont use consumables",
-    ["options.short.ShamanFlametongueRanked"]         = "Shaman: Use downranked Flametongue on offhand",
+    ["options.short.ShamanFlametongueRanked"]         = "Shaman: Use downranked Flametongue on mainhand",
 
     ["options.short.SomeoneIsDrinking"]               = "When someone is drinking...",
     ["options.convenience.SomeoneIsDrinking.Hide"]    = "Ignore the fact",
@@ -135,7 +135,9 @@ function englishModule:Translations()
     ["options.long.UseProfiles"]                      = "Use profiles based on whether the player is solo, in a group, raid or a battleground",
     ["options.long.UseRank"]                          = "Use spells with ranks",
     ["options.long.VisibleCategories"]                = "Show a category of buffs in the spells list (per character setting)",
-    ["options.long.ShamanFlametongueRanked"]          = "Shaman: For spellhancement shamans use highest rank Flametongue on main hand and down rank on the offhand. Use this when your weapon speeds are matching.",
+    ["options.long.ShamanFlametongueRanked"]          = "Shaman: For spellhancement shamans use downranked Flametongue "
+            .. "on main hand and max rank on the offhand. Use this when your weapon speeds are matching and you "
+            .. "have a spell power mainhand weapon.",
 
     ["tasklist.IgnoredBuffOn"]                        = "Ignored %s: %s", -- when a buff is not listed because a better buff exists
     ["task.target.Self"]                              = "Self", -- use instead of name when buffing self