+++ Src/LanguageRussian.lua
@@ -102,7 +102,7 @@ function russianModule:Translations()
     ["options.short.SomeoneIsDrinking"]         = "Когда кто-то в группе пьёт",
     ["options.short.UseRank"]                   = "Использовать заклинание с рангом",
     ["options.short.FadeWhenNothingToDo"]       = "Делать окно прозрачным, когда нет задач",
-    ["options.short.ShamanFlametongueRanked"]   = "Шаман: Использовать пониженный ранг языков пламени на левой руке",
+    ["options.short.ShamanFlametongueRanked"]   = "Шаман: Использовать пониженный ранг языков пламени на правой руке",
 
     ["options.long.FadeWhenNothingToDo"]        = "Установить прозрачность окна Бафомёта, если нечего делать",
     ["options.long.ActivateBomOnSpiritTap"]     = "Запретить Бафомёт, если 'Захват духа' жреца активен и мана меньше указанного процента",
@@ -117,7 +117,9 @@ function russianModule:Translations()
     ["options.long.Time3600"]                   = "Освежать баф продолжительностью <=60 мин, если осталось менее",
     ["options.long.Time60"]                     = "Освежать баф продолжительностью <=60 сек, если осталось менее",
     ["options.long.Time600"]                    = "Освежать баф продолжительностью <=10 мин, если осталось менее",
-    ["options.long.ShamanFlametongueRanked"]    = "Шаман: Для spellhancement шаманов использовать язык пламени максимального ранга на правом оружии, и на 1 ранг меньше на левом. Включайте эту опцию, когда скорость атаки обеих рук одинаковая.",
+    ["options.long.ShamanFlametongueRanked"]    = "Шаман: Для spellhancement шаманов использовать язык пламени "
+            .. "пониженного ранга на правом оружии, и максимального ранга на левом. "
+            .. "Включайте эту опцию, когда в правой руке у вас оружие с силой заклинаний.",
 
     ["options.short.MinBlessing"]               = "Большое благословение, если нуждаются более",
     ["options.short.MinBuff"]                   = "Групповой баф, если нуждаются более",